flask/ping_pong_app/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.30
numpy==1.20.3
pandas==1.2.5
